+++ R/spread2.R
@@ -84,9 +84,9 @@ if (getRversion() >= "3.1.0") {
 #'                       on during a spreading event. This will override an event that
 #'                       stops probabilistically via \code{spreadProb}, but forcing
 #'                       its last set of active cells to try again to find neighbours.
-#'                       It will try 10 times per event, before giving up.
-#'                       During those 10 times, it will try twice to "jump" up to
-#'                       4 cells outwards from each of the active cells.\cr
+#'                       It will try \code{maxRetriesPerID} times per event, before giving up.
+#'                       During those \code{maxRetriesPerID} times, it will try to "jump" up to
+#'                       4 cells outwards from each of the active cells, every 5 retries.\cr
 #'   \code{iterations} \tab This is a hard cap on the number of internal iterations to
 #'                          complete before returning the current state of the system
 #'                          as a \code{data.table}.\cr
@@ -187,6 +187,10 @@ if (getRversion() >= "3.1.0") {
 #'                   spread iteration will spread to \code{1, 2, ..., length(neighProbs)}
 #'                   neighbours, respectively. If this is used (i.e., something other than
 #'                   NA), \code{circle} and \code{returnDistances} will not work currently.
+#' @param maxRetriesPerID Only active if \code{exactSize} is used. This is the number of attempts
+#'                        that will be made per event ID, before abandoning, therefore completing
+#'                        the spread2 for that event with a size that is smaller than
+#'                        \code{exactSize}. Default 10 times.
 #'
 #' @param asymmetry     A numeric or \code{RasterLayer} indicating the ratio of the
 #'                      asymmetry to be used. i.e., 1 is no asymmetry; 2 means that the
@@ -329,6 +333,7 @@ if (getRversion() >= "3.1.0") {
 #' @importFrom checkmate qassert
 #' @importFrom data.table := alloc.col as.data.table copy data.table is.data.table
 #' @importFrom data.table rbindlist set setattr setcolorder setkeyv setnames uniqueN
+#' @importFrom fastmatch fmatch
 #' @importFrom ff ff
 #' @importFrom fpCompare %<=% %>>%
 #' @importFrom magrittr %>%
@@ -345,7 +350,7 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
                     spreadProb = 0.23, persistProb = NA_real_, asRaster = TRUE,
                     maxSize, exactSize, directions = 8L, iterations = 1e6L,
                     returnDistances = FALSE, returnDirections = FALSE,
-                    returnFrom = FALSE,
+                    returnFrom = FALSE, maxRetriesPerID = 10,
                     spreadProbRel = NA_real_, plot.it = FALSE, circle = FALSE,
                     asymmetry = NA_real_, asymmetryAngle = NA_real_,
                     allowOverlap = 0, neighProbs = NA_real_, skipChecks = FALSE) {
@@ -354,6 +359,7 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
   assertClass(landscape, "Raster")
   ncells <- ncell(landscape)
   numCols <- ncol(landscape)
+  anyNAneighProbs <- any(is.na(neighProbs))
   if (!skipChecks) {
     assert(
       checkNumeric(start, min.len = 0, max.len = ncells, lower = 1, upper = ncells),
@@ -435,7 +441,7 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
   usingAsymmetry <- !is.na(asymmetry)
 
   # This means that if an event can not spread any more, it will try 10 times, incl. 2 jumps
-  maxRetriesPerID <- 10
+  # maxRetriesPerID <- 10
 
   if (!is.numeric(start) & !is.data.table(start)) {
     if (is(start, "Raster")) {
@@ -742,7 +748,7 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
     set(dtPotential, NULL, "state", "successful")
 
     # Step 5 -- optional -- Algorithm neighProbs - uses a specific number of neighbours
-    if (!anyNA(neighProbs)) {
+    if (!anyNAneighProbs) {
       # numNeighs algorithm
       numNeighsByPixel <- unique(dtPotential, by = c("id", "from"))
       if (is.list(neighProbs)) {
@@ -792,12 +798,20 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
         if (NROW(dtPotential)) {
           # If it is a corner or has had pixels removed bc of duplicates,
           # it may not have enough neighbours
-          numNeighsByPixel <- numNeighsByPixel[dtPotential[, .N, by = c("id", "from")]]
-          set(numNeighsByPixel, NULL, "numNeighs",
-              pmin(numNeighsByPixel$N, numNeighsByPixel$numNeighs, na.rm = TRUE))
-          dtPotential <- dtPotential[numNeighsByPixel[dtPotential][,
-                                                                   .I[sample.int(length(numNeighs), size = numNeighs, prob = spreadProbRel)],
-                                                                   by = "from"]$V1]
+          set(numNeighsByPixel, NULL, c("to", "state"), NULL)
+          dt1 <- dtPotential[numNeighsByPixel, nomatch = 0]
+          dtPotential <- dt1[dt1[, .I[sample.int(.N, size = min(.N, numNeighs), prob = spreadProbRel)], by = c("id", "from")]$V1]
+
+          if (FALSE) { # old algorithm, replaced by 3 lines above May 30 2019, Eliot -- appears to be a bug below
+            #   the by = "from" should be c("id", "from") -- should sample 1 or more from each fire event, from each front line
+            numNeighsByPixel <- numNeighsByPixel[dtPotential[, .N, by = c("id", "from")]]
+            if (any(numNeighsByPixel$numNeighs > numNeighsByPixel$N))
+              set(numNeighsByPixel, NULL, "numNeighs",
+                  pmin(numNeighsByPixel$N, numNeighsByPixel$numNeighs, na.rm = TRUE))
+            dtPotential <- dtPotential[numNeighsByPixel[dtPotential][,
+                                                                      .I[sample.int(length(numNeighs), size = numNeighs, prob = spreadProbRel)],
+                                                                      by = "from"]$V1]
+          }
         }
         set(dtPotential, NULL, "spreadProbRel", NULL)
       }
@@ -847,7 +861,7 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
     spreadProbSuccess <- runifC(NROW(dtPotential)) <= actualSpreadProb
 
     # Step 8 - Remove duplicates & bind dt and dtPotential
-    if (anyNA(neighProbs)) {
+    if (anyNAneighProbs) {
       if (isTRUE(allowOverlap > 0) | is.na(allowOverlap) | !canUseAvailable) {
         # overlapping allowed
         dtPotential <- dtPotential[spreadProbSuccess]
@@ -944,17 +958,22 @@ spread2 <- function(landscape, start = ncell(landscape) / 2 - ncol(landscape) / 
       # Too small second
       if (!(anyNA(exactSize))) {
         # push those that are too small into "tooSmall"
-        currentSizeTooSmall <- clusterDT[tooBigByNCells < 0]
+        currentSizeTooSmall <- clusterDT[tooBigByNCells < 0, "initialPixels"]
+        # dtOrig <- copy(dt)
+        # csts <- copy(currentSizeTooSmall)
+        #dt <- copy(dtOrig)
+        #currentSizeTooSmall <- copy(csts)
         if (NROW(currentSizeTooSmall) > 0) {
           # successful means will become activeSource next iteration,
           # so they don't need any special treatment
           currentSizeTooSmall <- currentSizeTooSmall[
-            !dt[state %in% c("successful", "holding"), nomatch = 0]
+            !dt[dt$state %in% c("successful", "holding"), nomatch = 0]
             ]
+
         }
         # if the ones that are too small are unsuccessful, make them "tooSmall"
         set(dt, NULL, "ind", seq_len(NROW(dt)))
-        whTooSmall <- dt[!(state %in% c("successful", "inactive"))][
+        whTooSmall <- dt[!(dt$state %in% c("successful", "inactive"))][
           currentSizeTooSmall, nomatch = 0]$ind
         set(dt, NULL, "ind", NULL)
 
@@ -1113,9 +1132,8 @@ rbindlistDtDtpot <- function(dt, dtPotential, returnFrom, needDistance, dtPotent
 #' Internal helpers for \code{spread2}
 #'
 #' @inheritParams rbindlistDtDtpot
-#' @rdname spread2-internals
 #' @keywords internal
-#'
+#' @rdname spread2-internals
 reorderColsWDistance <- function(needDistance, dtPotential, dtPotentialColNames) {
   if (needDistance)
     setcolorder(dtPotential,
@@ -1133,7 +1151,6 @@ reorderColsWDistance <- function(needDistance, dtPotential, dtPotentialColNames)
 #'                             vector \code{NROW(dtPotential)}.
 #' @keywords internal
 #' @rdname spread2-internals
-#'
 angleQuality <- function(from, to, landscape, actualAsymmetryAngle) {
   from1 <- cbind(id = from, xyFromCell(landscape, from))
   to1 <- cbind(id = from, xyFromCell(landscape, to))