+++ tests/testthat/test-spread2.R
@@ -26,7 +26,7 @@ test_that("spread2 tests", {
     expect_true(all(out$active == 0))
   }
 
-  if (interactive()) print("testing maxSize")
+  if (interactive()) message("testing maxSize")
   maxSizes <- 2:3
   for (i in 1:20) {
     seed <- sample(1e6, 1)
@@ -37,7 +37,7 @@ test_that("spread2 tests", {
     expect_true(all(out[, .N, by = "initialPixels"]$N <= maxSizes[order(sams)]))
   }
 
-  if (interactive()) print("testing exactSize")
+  if (interactive()) message("testing exactSize")
   exactSizes <- c(5, 3.1)
   for (i in 1:20) {
     sams <- sample(innerCells, 2)
@@ -58,7 +58,7 @@ test_that("spread2 tests", {
     }
   }
 
-  if (interactive()) print("testing exactSize")
+  if (interactive()) message("testing exactSize")
   exactSizes <- c(5.01, 3.1, 4)
   for (i in 1:20) {
     sams <- sample(innerCells, length(exactSizes))
@@ -73,7 +73,7 @@ test_that("spread2 tests", {
     }
   }
 
-  if (interactive()) print("testing exact maxSize, can't be achieved, allow jumping")
+  if (interactive()) message("testing exact maxSize, can't be achieved, allow jumping")
   exactSizes <- c(154, 111, 134) # too big for landscape, can't achieve it --
                                  #  will hit max numRetries, and will try jumping
   for (i in 1:20) {
@@ -85,7 +85,7 @@ test_that("spread2 tests", {
     expect_true(all(out$numRetries == 11)) # current max
   }
 
-  if (interactive()) print("test circle = TRUE")
+  if (interactive()) message("test circle = TRUE")
   for (i in 1:20) {
     message(i)
     seed <- sample(1e6, 1)
@@ -129,7 +129,7 @@ test_that("spread2 tests", {
     }
   }
 
-  if (interactive()) print("compare spread2 circle with cir circle")
+  if (interactive()) message("compare spread2 circle with cir circle")
   cirOut <- data.table(
     cir(a, allowOverlap = TRUE, loci = sams, minRadius = 0, maxRadius = 15,
         returnDistances = TRUE, simplify = TRUE)
@@ -152,7 +152,7 @@ test_that("spread2 tests", {
   compare <- out[cirOut, on = c(initialPixels = "initialPixels", pixels = "indices")]
   expect_true(sum(abs(compare$dists - compare$distance)) %==% 0)
 
-  if (interactive()) print("Scales with number of starts, not maxSize of raster")
+  if (interactive()) message("Scales with number of starts, not maxSize of raster")
   set.seed(21)
   b <- raster(extent(0, 33000, 0, 33000), res = 1)
   sams <- sample(ncell(b), 2)
@@ -161,7 +161,7 @@ test_that("spread2 tests", {
   })
   expect_lt(st1[1], 1)
 
-  if (interactive()) print("test neighProbs")
+  if (interactive()) message("test neighProbs")
   maxSizes <- 14
   sp <- raster(a)
   spreadProbOptions <- 1:5
@@ -174,7 +174,7 @@ test_that("spread2 tests", {
   expect_true(uniqueN(out) == maxSizes * length(sams))
   expect_true(NROW(out) == maxSizes * length(sams))
 
-  if (interactive()) print("check variable lengths of neighProbs")
+  if (interactive()) message("check variable lengths of neighProbs")
   set.seed(29937)
   sams <- sample(innerCells, 2)
   for (i in 1:8) {
@@ -193,7 +193,7 @@ test_that("spread2 tests", {
   }
 
   if (interactive()) {
-    print(
+    message(
       paste(
         "Test that when using neighProbs & a Raster of spreadProbs,",
         "the spreadProb raster is followed probabilistically.",
@@ -224,7 +224,7 @@ test_that("spread2 tests", {
     expect_false(cht$p.value > 0.05) ## TODO: is this valid/correct test?
   }
 
-  print("Scales with number of starts, not maxSize of raster")
+  message("Scales with number of starts, not maxSize of raster")
   set.seed(21)
   b <- raster(extent(0, 10, 0, 10), res = 1)
   bProb <- gaussMap(b, speedup = 1)
@@ -238,7 +238,7 @@ test_that("spread2 tests", {
   if (interactive()) out
 
   if (interactive())
-    print("check wide range of spreadProbs and that it makes a RasterLayer")
+    message("check wide range of spreadProbs and that it makes a RasterLayer")
   set.seed(654)
   rasts <- list()
   for (i in 1:20) {
@@ -251,7 +251,7 @@ test_that("spread2 tests", {
   }
 
   if (interactive())
-    print("testing iterative calling of spread2")
+    message("testing iterative calling of spread2")
   set.seed(299)
   sams <- sample(innerCells, 2)
   set.seed(299)
@@ -267,7 +267,7 @@ test_that("spread2 tests", {
   expect_equivalent(out2, out)
 
   if (interactive())
-    print("testing iterative calling of spread2, but asRaster = TRUE")
+    message("testing iterative calling of spread2, but asRaster = TRUE")
   set.seed(299)
   sams <- sample(innerCells, 2)
   set.seed(299)
@@ -280,7 +280,7 @@ test_that("spread2 tests", {
   expect_true(identical(out, attr(out1, "pixel")))
 
   if (interactive())
-    print("testing iterative with maxSize")
+    message("testing iterative with maxSize")
   set.seed(299)
   seed <- sample(1e6, 1)
   set.seed(seed)
@@ -294,7 +294,7 @@ test_that("spread2 tests", {
   }
 
   if (interactive())
-    print("testing iterative with maxSize -- where needRetry occurs")
+    message("testing iterative with maxSize -- where needRetry occurs")
   set.seed(299)
   sams <- sample(innerCells, 2)
   exactSizes <- 60:61
@@ -354,251 +354,6 @@ test_that("spread2 tests", {
   }
   expect_false(identical(data.table(out2), data.table(out)))
 
-  ##############################################################
-  skip("benchmarking spread2")
-  exactSizes <- 60:61
-  microbenchmark(
-    times = 10,
-    a = {
-      out <- spread2(a, start = sams, spreadProb = 0.225, iterations = 1,
-                     exactSize = exactSizes, asRaster = FALSE)
-      for (i in 1:25) {
-        out <- spread2(a, start = out, spreadProb = 0.225, iterations = 1,
-                       exactSize = exactSizes, asRaster = FALSE)
-      }
-    },
-    b = {
-      out2 <- spread2(a, start = sams, spreadProb = 0.225, iterations = 1,
-                      exactSize = exactSizes, asRaster = FALSE)
-      for (i in 1:25) {
-        attr(out2, "spreadState") <- NULL
-        out2 <- spread2(a, start = out2, spreadProb = 0.225, iterations = 1,
-                        exactSize = exactSizes, asRaster = FALSE)
-      }
-    }
-  )
-
-  a <- raster(extent(0, 1000, 0, 1000), res = 1)
-  set.seed(123)
-  sams <- sample(innerCells, 30)
-  set.seed(123)
-  profvis::profvis({
-    out <- spread2(a, start = sams, spreadProb = 0.235, asRaster = FALSE)
-  })
-  set.seed(123)
-  profvis::profvis({
-    out <- spread2(a, start = sams, spreadProb = 0.235, asRaster = FALSE, allowOverlap = TRUE)
-  })
-
-  set.seed(123)
-  microbenchmark(times = 30, {
-    out1 <- spread2(a, start = sams, spreadProb = 0.235, asRaster = FALSE)
-  },
-  b = {
-    out2 <- spread(a, loci = sams, spreadProb = 0.235, id = TRUE)
-  },
-  c = {
-    out2 <- spread(a, loci = sams, spreadProb = 0.235, id = TRUE, lowMemory = TRUE)
-  })
-  set.seed(123)
-  profvis::profvis({
-    out <- spread2(a, start = sams, spreadProb = 0.235, asRaster = FALSE, allowOverlap = TRUE)
-  })
-
-  ######## Benchmarking ##########
-  iterativeFun <- function(a, skipChecks, n, sp) {
-    sams <- sample(innerCells, n)
-    out <- spread2(a, iterations = 1, start = sams, skipChecks = skipChecks,
-                   asRaster = FALSE, spreadProb = sp)
-    stillActive <- TRUE
-    while (stillActive) {
-      stillActive <- any(out$state == "activeSource")
-      out <- spread2(a, spreadProb = sp, iterations = 1, start = out,
-                     asRaster = FALSE, skipChecks = skipChecks)
-    }
-    out
-  }
-
-  nonIterativeFun <- function(a, skipChecks, n, sp) {
-    sams <- sample(innerCells, n)
-    out <- spread2(a, start = sams, asRaster = FALSE, skipChecks = skipChecks, spreadProb = sp)
-    out
-  }
-
-  origSpread <- function(a, quick, n, sp) {
-    sams <- sample(innerCells, n)
-    out <- spread(a, spreadProb = sp, loci = sams, id = TRUE, returnIndices = TRUE,
-                  quick = quick)
-    out
-  }
-
-  origSpreadIterations <- function(a, quick, n, sp) {
-    sams <- sample(innerCells, n)
-    out <- spread(a, spreadProb = sp, iterations = 1, loci = sams, returnIndices = TRUE)
-    stillActive <- TRUE
-    while (stillActive) {
-      stillActive <- any(out$active)
-      out <- spread(a, spreadProb = sp, iterations = 1, spreadState = out,
-                    returnIndices = TRUE, quick = quick)
-    }
-    out
-  }
-
-  n <- 2
-  ras <- raster(extent(0, 160, 0, 160), res = 1)
-  n <- 2000
-  ras <- raster(extent(0, 6000, 0, 6000), res = 1)
-  sp <- 0.225
-  b <- raster(ras)
-  b[] <- 1
-  bb <- focal(b, matrix(1 / 9, nrow = 3, ncol = 3), fun = sum, pad = TRUE, padValue = 0)
-  innerCells <- which(bb[] %==% 1)
-
-  microbenchmark(
-    times = 3,
-    iterativeFun(ras, TRUE, n, sp),
-    nonIterativeFun(ras, TRUE, n, sp),
-    origSpread(ras, TRUE, n, sp),
-    origSpreadIterations(ras, TRUE, n, sp)
-  )
-  #         iterativeFun(ras, TRUE, n, sp) 2.581752 2.667209 2.727886 2.752666 2.800953 2.849239
-  #      nonIterativeFun(ras, TRUE, n, sp) 2.009914 2.268422 2.486540 2.526930 2.724854 2.922777
-  #           origSpread(ras, TRUE, n, sp) 1.003594 1.043065 1.085511 1.082536 1.126470 1.170404
-  # origSpreadIterations(ras, TRUE, n, sp) 7.267927 7.630435 7.908982 7.992943 8.229510 8.46607
-#
-  # without "skipChecks"
-  microbenchmark(
-    times = 100,
-    iterativeFun(ras, FALSE, n, sp),
-    nonIterativeFun(ras, FALSE, n, sp),
-    origSpread(ras, FALSE, n, sp),
-    origSpreadIterations(ras, FALSE, n, sp)
-  )
-  ## Unit: milliseconds
-  ##                                    expr       min        lq      mean   median        uq       max neval
-  ##         iterativeFun(ras, FALSE, n, sp)  3.096979 12.642477  73.02248 35.17520  91.10528  764.4073   300
-  ##      nonIterativeFun(ras, FALSE, n, sp)  1.509484  6.555565  31.18444 14.91066  42.78317  158.5237   300
-  ##           origSpread(ras, FALSE, n, sp)  5.154006  7.555631  14.87158 11.49005  17.50599  231.5487   300
-  ## origSpreadIterations(ras, FALSE, n, sp) 10.754669 51.524368 141.48620 93.61996 169.10808 2110.2683   300
-  ##
-  profvis::profvis({
-    set.seed(3451)
-    for (i in 1:8)
-      iterativeFun(ras, TRUE, n, sp = sp)
-  })
-  profvis::profvis({
-    nonIterativeFun()
-  })
-
-  library(raster)
-  library(data.table)
-  library(fpCompare)
-  n <- 2
-  ras <- raster(extent(0, 160, 0, 160), res = 1)
-  sp <- 0.225
-  b <- raster(ras)
-  b[] <- 1
-  bb <-
-    focal(
-      b,
-      matrix(1 / 9, nrow = 3, ncol = 3),
-      fun = sum,
-      pad = TRUE,
-      padValue = 0
-    )
-  innerCells <- which(bb[] %==% 1)
-
-  iterativeNeigh <- function(a, skipChecks, n, sp, exactSizes,
-                           neighProbs, ...) {
-    sams <- sample(innerCells, n)
-    out <- spread2(a, iterations = 1, spreadProb = sp, start = sort(sams),
-                   neighProbs = neighProbs, exactSize = exactSizes,
-                   skipChecks = skipChecks, asRaster = FALSE, ...)
-    stillActive <- TRUE
-    while (stillActive) {
-      out <- spread2(a, iterations = 1, spreadProb = sp, start = out,
-                     neighProbs = neighProbs, exactSize = exactSizes,
-                     skipChecks = skipChecks, asRaster = FALSE, ...)
-      stillActive <- any(length(attr(out, "spreadState")$whActive) |
-                           length(attr(out, "spreadState")$whNeedRetry))
-    }
-    out
-  }
-
-  origSpreadIterationsNeighs <- function(a, quick, n, sp, neighProbs, exactSize) {
-    sams <- sample(innerCells, n)
-    out <- spread(a, spreadProb = sp, neighProbs = neighProbs, iterations = 1,
-                  loci = sams, exactSizes = TRUE, returnIndices = TRUE, maxSize = exactSize)
-    stillActive <- TRUE
-    while (stillActive) {
-      stillActive <- any(out$active)
-      out <- spread(a, spreadProb = sp, iterations = 1, neighProbs = neighProbs,
-                    spreadState = out, maxSize = exactSize, exactSizes = TRUE,
-                    returnIndices = TRUE, quick = quick)
-    }
-    out
-  }
-
-  sp <- randomPolygons(ras, numTypes = 35)
-  sp[] <- (sp[] %% 5 + 1) / 10
-  exactSizes <- c(123, 2240)
-  neighProbs <- c(0.5, 0.3, 0.2)
-
-  microbenchmark(
-    times = 9,
-    a = {
-      iterativeNeigh(ras, TRUE, length(exactSizes), sp = sp, exactSize = exactSizes,
-                     neighProbs = neighProbs)
-    },
-    b = {
-      origSpreadIterationsNeighs(ras, TRUE, length(exactSizes), sp = sp,
-                               exactSize = exactSizes, neighProbs = neighProbs)
-    }
-  )
-
-  profvis::profvis({
-    for (i in 1:5) {
-      iterativeNeigh(ras, TRUE, length(exactSizes), sp = sp, exactSize = exactSizes,
-                     neighProbs = neighProbs)
-    }
-  })
-
-  profvis::profvis({
-    for (i in 1:5) {
-      origSpreadIterationsNeighs(ras, TRUE, length(exactSizes), sp = sp,
-                                 exactSize = exactSizes, neighProbs = neighProbs)
-    }
-  })
-
-  dev()
-  iterativeNeigh(ras, TRUE, length(exactSizes), sp = sp,  exactSize = exactSizes, plot.it = TRUE)
-  iterativeNeigh(ras, TRUE, length(exactSizes), sp = sp)#,  plot.it = TRUE)
-  iterativeNeigh(ras, TRUE, length(exactSizes), sp = sp)#,  plot.it = TRUE)
-
-
-  # compare original spread and spread2 -- seems pretty dead on
-  nn <- 1000
-  outNew <- out <- numeric(nn)
-  for (i in 1:nn) {
-    outNew[i] <- NROW(nonIterativeFun(ras, TRUE, n, sp))
-    out[i] <- NROW(origSpread(ras, TRUE, n, sp))
-  }
-
-  out <- data.table(x = out)
-  outNew <- data.table(x = outNew)
-  mean(out$x)
-  mean(outNew$x)
-  sd(out$x)
-  sd(outNew$x)
-
-  n <- 5
-  ras <- raster(extent(0, 1000, 0, 1000), res = 1)
-  sp <- 0.295
-  set.seed(123)
-  microbenchmark(
-    times = 100,
-    nonIterativeFun(ras, TRUE, n, sp)
-  )
 })
 
 test_that("spread2 tests -- asymmetry", {