+++ NEWS.md
@@ -1,5 +1,18 @@
 Known issues: https://github.com/PredictiveEcology/SpaDES.tools/issues
 
+version 0.3.2.9000
+=============
+
+## Dependency changes
+* Added `backports` for R-oldrel support
+* remove `profvis` and `microbenchmark` from Suggests: all benchmarking is removed from automated tests
+* Requires R >= 3.5
+
+## Bugfixes
+
+## Enhancements
+* `spread2` speed enhancements
+
 version 0.3.2
 =============